Rft: move connection from client to transport mod
